Fix a bug of concurrently executing UPDATE/DELETE with VACUUM on a partition ao table
The problem is ERROR "tuple concurrently updated" occurs when running
delete qeury during ETL as below error. When we concurrently execute an
UPDATE/DELETE statement and a VACUUM statement on a partition ao table.
UPDATE/DELETE statement is likely to get an out-of-date snapshot after
the VACUUM has been processed on one of the leaf partition table.
So we fix it by acquiring locks before executing. After we get locks, we
dispatch snapshot. So this problem will not happen.

Co-authored-by: Rui Wang wangru@vmware.com wraymo